requirements.txt
Django>=4.0
djangorestframework>=3.13.1
sqlparse==0.5.3
Django>=4.0
djangorestframework>=3.13.1
